+++ ui/components/app/modals/contract-details-modal/index.scss
@@ -12,15 +12,6 @@
         background: none;
       }
 
-      &__collection {
-        border-radius: 50%;
-        width: 24px;
-        height: 24px;
-        padding: 2px;
-        margin: 16px;
-        background: var(--color-overlay-alternative);
-      }
-
       &__identicon-for-unknown-contact {
         margin: 16px;
       }
