+++ ui/components/ui/nft-collection-image/index.js
@@ -1,0 +1,1 @@
+export { default } from './nft-collection-image';
