+++ ui/components/ui/nft-collection-image/index.scss
@@ -1,0 +1,7 @@
+.collection-image-alt {
+  border-radius: 50%;
+  width: 24px;
+  height: 24px;
+  padding: 2px;
+  background: var(--color-overlay-alternative);
+}
