+++ ui/components/ui/ui-components.scss
@@ -63,3 +63,5 @@
 @import 'deprecated-test-networks/index.scss';
 @import 'contract-token-values/index.scss';
 @import 'nft-info/index.scss';
+@import 'nft-collection-image/index';
+
